refactoring bollinger-bands file into bot-skeleton
