+++ packages/bot-skeleton/src/services/tradeEngine/indicators/simple-moving-average.js
